requirements.txt
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
Flask==2.1.3
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pip==22.3.1
python-dotenv==0.21.0
setuptools==65.6.3
SQLAlchemy==1.4.39
visitor==0.1.3
Werkzeug==2.0.3
wheel==0.38.4
WTForms==3.0.1
